Add debug logging to types
